+++ aiobastion/accounts.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 import asyncio
 import re
-from typing import List, Union, AsyncIterator, AsyncGenerator
+from typing import List, Union, AsyncIterator
+
 import aiohttp
 
 from .config import validate_ip, flatten
@@ -338,7 +339,7 @@ class Account:
                                           reconcile_account: PrivilegedAccount):
         """
         | This function links the account (or the list of accounts) to the given reconcile account
-        | ⚠️ The "reconcile" Account is supposed to have an index of 3
+        | ⚠️ The "reconcile" Account index is default to 3
 
         :param account: a PrivilegedAccount object or a list of PrivilegedAccount objects
         :type account: PrivilegedAccount, list
@@ -346,13 +347,13 @@ class Account:
         :return: A boolean that indicates if the operation was successful.
         :raises CyberarkException: If link failed
         """
-        return await self.link_account(account, reconcile_account, 3)
+        return await self.link_account(account, reconcile_account, self.epv.RECONCILE_ACCOUNT_INDEX)
 
     async def link_logon_account(self, account: Union[PrivilegedAccount, List[PrivilegedAccount]],
                                  logon_account: PrivilegedAccount):
         """
         | This function links the account (or the list of accounts) to the given logon account
-        | ⚠️ The "logon" Account is supposed to have an index of 2
+        | ⚠️ The "logon" Account index is default to 2, you can change it by setting custom:LOGON_ACCOUNT_INDEX in the config
 
         :param account: a PrivilegedAccount object or a list of PrivilegedAccount objects
         :type account: PrivilegedAccount, list
@@ -361,7 +362,7 @@ class Account:
         :raises CyberarkException: If link failed
         """
         #TODO check the index of logon account at platform level !
-        return await self.link_account(account, logon_account, 2)
+        return await self.link_account(account, logon_account, self.epv.LOGON_ACCOUNT_INDEX)
 
     async def link_reconcile_account_by_address(self, acc_username, rec_acc_username, address):
         """ This function links the account with the given username and address to the reconciliation account with
@@ -396,7 +397,7 @@ class Account:
     async def remove_reconcile_account(self, account: Union[PrivilegedAccount, List[PrivilegedAccount]]):
         """
         | This function unlinks the reconciliation account of the given account (or the list of accounts)
-        | ⚠️ The "reconcile" Account is supposed to have an index of 3
+        | ⚠️ The "reconcile" Account index is default to 3
 
 
         :param account: a PrivilegedAccount object or a list of PrivilegedAccount objects
@@ -404,23 +405,24 @@ class Account:
         :return: A boolean that indicates if the operation was successful.
         :raises CyberarkException: If link failed:
         """
-        return await self.unlink_account(account, 3)
+        return await self.unlink_account(account, self.epv.RECONCILE_ACCOUNT_INDEX)
 
     async def remove_logon_account(self, account: Union[PrivilegedAccount, List[PrivilegedAccount]]):
         """
         | This function unlinks the logon account of the given account (or the list of accounts)
-        | ⚠️ The "logon" Account is supposed to have an index of 2
+        | ⚠️ The "logon" Account index is default to 2, you can change it by setting custom:LOGON_ACCOUNT_INDEX in the config
 
         :param account: a PrivilegedAccount object or a list of PrivilegedAccount objects
         :type account: PrivilegedAccount, list
         :return: A boolean that indicates if the operation was successful.
         :raises CyberarkException: If link failed:
         """
-        return await self.unlink_account(account, 2)
+        return await self.unlink_account(account, self.epv.LOGON_ACCOUNT_INDEX)
 
     async def unlink_account(self, account: Union[PrivilegedAccount, List[PrivilegedAccount]],
                              extra_password_index: int):
         """ This function unlinks the account of the given account (or the list of accounts)
+        | ⚠️ Double check the linked account index on your platform.
 
         :param account: a PrivilegedAccount object or a list of PrivilegedAccount objects
         :type account: PrivilegedAccount, list
@@ -682,7 +684,7 @@ class Account:
             "has_next_page": has_next_page
         }
 
-    async def connect_using_PSM(self, account, connection_component):
+    async def connect_using_PSM(self, account, connection_component, reason: str = ""):
         """ This function returns a file content (bytes) which is the equivalent RDP file of the “Connect” button
 
         For example::
@@ -707,6 +709,7 @@ class Account:
 
         :param account: PrivilegedAccount or account_id
         :param connection_component: the connection component to connect with
+        :param reason: the reason that is required to request access to this account
         :return: file_content
         :raises CyberarkAPIException:  if an error occured
         """
@@ -714,6 +717,7 @@ class Account:
         url, head = self.epv.get_url(f"API/Accounts/{account_id}/PSMConnect")
         head["Accept"] = 'RDP'
         body = {"ConnectionComponent": connection_component}
+        if reason: body["Reason"] = reason
 
         async with aiohttp.ClientSession(headers=head) as session:
             async with session.post(url, json=body, **self.epv.request_params) as req:
@@ -919,12 +923,13 @@ class Account:
         else:
             raise AiobastionException("There is no CPM version for this account")
 
-    async def get_secret_version(self, account: PrivilegedAccount, version: int):
+    async def get_secret_version(self, account: PrivilegedAccount, version: int, reason: str = None):
         """
         Get the version of a password
 
         :param account: a PrivilegedAccount object
         :param version: the version ID (that you can find with get_secret_versions). The higher is the most recent.
+        :param reason: The reason that is required to retrieve the password
         :return: the secret
         :raises CyberarkException: if the version was not found
         """
@@ -932,26 +937,31 @@ class Account:
             raise AiobastionException("The version must be a non-zero natural integer")
 
         data = {"Version": version}
+        if reason: data["Reason"] = reason
         account_id = await self.get_account_id(account)
 
         url = f"API/Accounts/{account_id}/Password/Retrieve"
 
         return await self.epv.handle_request("post", url, data=data)
 
-    async def get_password(self, account: Union[PrivilegedAccount, str, List[PrivilegedAccount], List[str]]):
+    async def get_password(self, account: Union[PrivilegedAccount, str, List[PrivilegedAccount], List[str]], reason: str = None):
         """
         | Retrieve the password of an address
         | ✅ Use get_secret instead if you want to retrieve password or ssh_key
 
         :param account: a PrivilegedAccount object or a list of PrivilegedAccount objects
         :type account: PrivilegedAccount, list
+        :param reason: The reason that is required to retrieve the password
         :return: Account password value  (or list of passwords)
         :raises CyberarkException: If retrieve failed
         """
+        data = {}
+        if reason: data = {"Reason": reason}
         return await self._handle_acc_id_list(
             "post",
             lambda account_id: f"API/Accounts/{account_id}/Password/Retrieve",
-            await self.get_account_id(account)
+            await self.get_account_id(account),
+            data = data
         )
 
 
@@ -971,17 +981,21 @@ class Account:
             await self.get_account_id(account)
         )
 
-    async def get_secret_versions(self, account: Union[PrivilegedAccount, str, List[PrivilegedAccount], List[str]]):
+    async def get_secret_versions(self, account: Union[PrivilegedAccount, str, List[PrivilegedAccount], List[str]], reason: str = None):
         """
         Retrieve the secret versions
 
         :param account: Privileged Account or address id
+        :param reason: The reason that is required to retrieve the password
         :return: Account password value
         """
+        data = {}
+        if reason: data = {"Reason": reason}
         versions = await self._handle_acc_id_list(
             "get",
             lambda account_id: f"API/Accounts/{account_id}/Secret/Versions/",
-            await self.get_account_id(account)
+            await self.get_account_id(account),
+            data = data
         )
 
         if isinstance(versions, list):