+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-02-03
+### Changes
+- Add "get_safe_details" method
+- Add support for "custom" configs to override the default logon and reconcile account index
+- Add support to retain cookies during login, and use for subsequent API calls for load-balanced PVWAs.
+
+## [1.0.0] - 2024-01-26
+### Changes
+- Adding some debug information
+- Code refactoring
+
+## [0.0.31] - 2024-01-15
+### Changes
+- Add support to provide "Reason" when retrieving passwords
+- Update documentation for testing
+
+## [0.0.30] - 2024-16-01
+### Changes
+- Adding platform deactivation
+
+### Bugfixes
+- Typo in platform suppression
+- request_params not initialized when checking token
+
 ## [0.0.29] - 2024-11-01
 ### Changes
 - Adding AIM related functions get_password_aim and get_secret_aim